Remove base64 encoding of keys
I'm going to remove base64 encoding, since I want to focus on
correctness at the moment, and using base64 makes it harder to inspect
the keys. Instead we will print the keys as decimal numbers, which can
be read later by the encoder/decoder. This feature is already provided
by Boost.